chore: extract types file
