+++ src/sfz/sfz-sampler.ts
@@ -1,6 +1,6 @@
 import { toMidi } from "../player/midi";
 import { Player } from "../player/player";
-import { SampleStart, SampleStop } from "../player/sample-player";
+import { SampleStart, SampleStop } from "../player/types";
 import { HttpStorage, Storage } from "../storage";
 import { SfzInstrument } from "./sfz-kits";
 import { loadSfzBuffers, loadSfzInstrument } from "./sfz-load";
